+++ apps/eaas-interface/src/api/embalm.ts
@@ -1,0 +1,13 @@
+import axios from "axios";
+import { SendEncryptedPayloadParams } from "../../../common/types";
+import { handleApiError } from "./utils";
+
+export async function sendPayload(params: SendEncryptedPayloadParams) {
+  try {
+    const res = await axios.post(`${import.meta.env.VITE_API_BASE_URL}/embalm/send-payload`, params);
+    return res.status === 200;
+  } catch (error) {
+    handleApiError(error);
+    return false;
+  }
+}
